+++ src/game/magic/spells/SpellsLvl07.cpp
@@ -558,7 +558,7 @@ static Vec3f GetChestPos(EntityHandle num) {
 	}
 	
 	if(VertexGroupId chest = EERIE_OBJECT_GetGroup(io->obj, "chest")) {
-		return io->obj->vertexWorldPositions[io->obj->grouplist[size_t(chest)].origin].v;
+		return io->obj->vertexWorldPositions[io->obj->grouplist[chest].origin].v;
 	}
 	
 	return io->pos + Vec3f(0.f, -120.f, 0.f);
@@ -572,7 +572,7 @@ void LightningStrikeSpell::Update() {
 	Entity * caster = entities.get(m_caster);
 	if(caster) {
 		if(VertexGroupId chest = EERIE_OBJECT_GetGroup(caster->obj, "chest")) {
-			m_caster_pos = caster->obj->vertexWorldPositions[caster->obj->grouplist[size_t(chest)].origin].v;
+			m_caster_pos = caster->obj->vertexWorldPositions[caster->obj->grouplist[chest].origin].v;
 		} else {
 			m_caster_pos = caster->pos;
 		}
